backend/pyproject.toml
[project]
name = "salon-os-backend"
version = "0.1.0"
description = "SalonOS Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "psycopg[binary]>=3.1.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib==1.7.4",
    "bcrypt>=4.0.1",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "rq==1.15.1",
    "apscheduler==3.10.4",
    "python-ulid>=2.2.0",
    "pytz>=2023.3",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.1",
    "black==23.11.0",
    "ruff==0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
ignore = ["E501"]
